+++ src/main/scripts/org/reaktivity/specification/nukleus/http/streams/rfc7230/flow.control/request.with.content.fragmented.by.target.window/server/nukleus.rpt
@@ -76,17 +76,23 @@ write flush
 
 read [0x40 0x00 0x00 0x02]
 read ${newTargetInputId}
-read [0x00 0x02 0x00 0x00]
+read [0x03 0x00 0x00 0x00]
 
 write [0x00 0x00 0x00 0x02]
 write ${newTargetInputId}
-write [0x00 0x0b]
-write "hello world"
+write [0x00 0x03]
+write "hel"
 write flush
 
 read [0x40 0x00 0x00 0x02]
 read ${newTargetInputId}
-read [0x0b 0x00 0x00 0x00]
+read [0x08 0x00 0x00 0x00]
+
+write [0x00 0x00 0x00 0x02]
+write ${newTargetInputId}
+write [0x00 0x08]
+write "lo world"
+write flush
 
 write [0x00 0x00 0x00 0x03]
 write ${newTargetInputId}
