+++ src/main/scripts/org/reaktivity/specification/nukleus/http/streams/rfc7230/flow.control/request.with.fragmented.content.flow.controlled.by.target/server/nukleus.rpt
@@ -1,0 +1,174 @@
+#
+# Copyright 2016-2017 The Reaktivity Project
+#
+# The Reaktivity Project licenses this file to you under the Apache License,
+# version 2.0 (the "License"); you may not use this file except in compliance
+# with the License. You may obtain a copy of the License at:
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+
+property newTargetInputRef ${nuklei:newReferenceId()} # external scope
+
+property nuklei ${nuklei:directory("target/nukleus-itests").streamsCapacity(1024 * 1024, 64 * 1024)}
+
+property sourceInput ${nuklei.streams("http", "source")}
+property targetInput ${nuklei.streams("target", "http#source")}
+property targetOutputEst ${nuklei.streams("http", "target")}
+property sourceOutputEst ${nuklei.streams("source", "http#target")}
+
+property newTargetInputId ${nuklei:newStreamId()}
+property newSourceOutputEstId ${nuklei:newStreamId()}
+property newTargetCorrelationId ${nuklei:newCorrelationId()}
+
+connect await ROUTED_INPUT
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(sourceInput.buffer)}
+        option writer ${agrona:oneToOneWriter(sourceInput.throttle)}
+connected
+
+read [0x00 0x00 0x00 0x01]
+read ([0..8]:sourceInputId)
+read ([0..8]:sourceInputRef)
+read ([0..8]:sourceCorrelationId)
+
+write [0x40 0x00 0x00 0x02]
+write ${sourceInputId}
+write [0x00 0x20 0x00 0x00]
+write flush
+
+read [0x00 0x00 0x00 0x02]
+read ${sourceInputId}
+read [0x00 0x43]
+read "POST / HTTP/1.1\r\n"
+read "Host: localhost:8080\r\n"
+read "Content-Length: 11\r\n"
+read "\r\n"
+read "hello "
+
+read [0x00 0x00 0x00 0x02]
+read ${newSourceInputId}
+read [0x00 0x05]
+read "world"
+
+write notify HTTP_REQUEST_DECODED
+
+write [0x40 0x00 0x00 0x02]
+write ${sourceInputId}
+write [0x08 0x00 0x00 0x00]
+write flush
+
+close
+closed
+
+connect await HTTP_REQUEST_DECODED
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(targetInput.throttle)}
+        option writer ${agrona:oneToOneWriter(targetInput.buffer)}
+connected
+
+write [0x00 0x00 0x00 0x01]
+write ${newTargetInputId}
+write ${newTargetInputRef}
+write ${newTargetCorrelationId}
+write [0x00] [0x07] ":scheme" [0x04] "http"
+write [0x00] [0x07] ":method" [0x04] "POST"
+write [0x00] [0x05] ":path" [0x01] "/"
+write [0x00] [0x0a] ":authority" [0x0e] "localhost:8080"
+write [0x00] [0x0e] "content-length" [0x02] "11"
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newTargetInputId}
+read [0x06 0x00 0x00 0x00]
+
+write [0x00 0x00 0x00 0x02]
+write ${newTargetInputId}
+write [0x00 0x06]
+write "hello "
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newTargetInputId}
+read [0x03 0x00 0x00 0x00]
+
+write [0x00 0x00 0x00 0x02]
+write ${newTargetInputId}
+write [0x00 0x03]
+write "wor"
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newTargetInputId}
+read [0x02 0x00 0x00 0x00]
+
+write [0x00 0x00 0x00 0x02]
+write ${newTargetInputId}
+write [0x00 0x02]
+write "ld"
+write flush
+
+write [0x00 0x00 0x00 0x03]
+write ${newTargetInputId}
+write flush
+
+close
+closed
+
+connect await ROUTED_OUTPUT
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(targetOutputEst.buffer)}
+        option writer ${agrona:oneToOneWriter(targetOutputEst.throttle)}
+connected
+
+read [0x00 0x00 0x00 0x01]
+read ([0..8]:targetOutputEstId)
+read [0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00]
+read ${newTargetCorrelationId}
+read [0x00] [0x07] ":status" [0x03] "200"
+
+read notify HTTP_RESPONSE_RECEIVED
+
+write [0x40 0x00 0x00 0x02]
+write ${targetOutputEstId}
+write [0x00 0x02 0x00 0x00]
+write flush
+
+close
+closed
+
+connect await HTTP_RESPONSE_RECEIVED
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(sourceOutputEst.throttle)}
+        option writer ${agrona:oneToOneWriter(sourceOutputEst.buffer)}
+connected
+
+write [0x00 0x00 0x00 0x01]
+write ${newSourceOutputEstId}
+write [0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00]
+write ${sourceCorrelationId}
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newSourceOutputEstId}
+read [0..4]
+
+write [0x00 0x00 0x00 0x02]
+write ${newSourceOutputEstId}
+write [0x00 0x13]
+write "HTTP/1.1 200 OK\r\n"
+write "\r\n"
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newSourceOutputEstId}
+read [0..4]
+
+close
+closed
