+++ src/main/scripts/org/reaktivity/specification/nukleus/http/streams/rfc7230/flow.control/request.with.fragmented.content.flow.controlled.by.target/server/source.rpt
@@ -1,0 +1,90 @@
+#
+# Copyright 2016-2017 The Reaktivity Project
+#
+# The Reaktivity Project licenses this file to you under the Apache License,
+# version 2.0 (the "License"); you may not use this file except in compliance
+# with the License. You may obtain a copy of the License at:
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+
+property newSourceInputRef ${nuklei:newReferenceId()} # external scope
+
+property nuklei ${nuklei:directory("target/nukleus-itests").streamsCapacity(1024 * 1024, 64 * 1024)}
+
+property sourceInput ${nuklei.streams("http", "source")}
+property sourceOutputEst ${nuklei.streams("source", "http#target")}
+
+property newSourceInputId ${nuklei:newStreamId()}
+property newSourceCorrelationId ${nuklei:newCorrelationId()}
+
+connect await ROUTED_INPUT
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(sourceInput.throttle)}
+        option writer ${agrona:oneToOneWriter(sourceInput.buffer)}
+connected
+
+write [0x00 0x00 0x00 0x01]
+write ${newSourceInputId}
+write ${newSourceInputRef}
+write ${newSourceCorrelationId}
+write flush
+
+read [0x40 0x00 0x00 0x02]
+read ${newSourceInputId}
+read [0..4]
+
+write [0x00 0x00 0x00 0x02]
+write ${newSourceInputId}
+write [0x00 0x43]
+write "POST / HTTP/1.1\r\n"
+write "Host: localhost:8080\r\n"
+write "Content-Length: 11\r\n"
+write "\r\n"
+write "hello "
+write flush
+
+write [0x00 0x00 0x00 0x02]
+write ${newSourceInputId}
+write [0x00 0x05]
+write "world"
+write flush
+
+close
+closed
+
+connect await ROUTED_OUTPUT
+        agrona://stream/bidirectional
+        option reader ${agrona:oneToOneReader(sourceOutputEst.buffer)}
+        option writer ${agrona:oneToOneWriter(sourceOutputEst.throttle)}
+connected
+
+read [0x00 0x00 0x00 0x01]
+read ([0..8]:sourceOutputEstId)
+read [0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00]
+read ${newSourceCorrelationId}
+
+write [0x40 0x00 0x00 0x02]
+write ${sourceOutputEstId}
+write [0x00 0x02 0x00 0x00]
+write flush
+
+read [0x00 0x00 0x00 0x02]
+read ${sourceOutputEstId}
+read [0x00 0x13]
+read "HTTP/1.1 200 OK\r\n"
+read "\r\n"
+
+write [0x40 0x00 0x00 0x02]
+write ${sourceOutputEstId}
+write [0x13 0x00 0x00 0x00]
+write flush
+
+close
+closed
