+++ src/main/scripts/org/reaktivity/specification/nukleus/http/streams/rfc7230/flow.control/request.with.fragmented.content.flow.controlled.by.target/server/target.rpt
@@ -21,6 +21,8 @@ property targetOutputEst ${nuklei.streams("http", "target")}
 
 property newTargetOutputEstId ${nuklei:newStreamId()}
 
+property targetInputInitialWindow [0x00 0x20 0x00 0x00]
+
 connect await ROUTED_INPUT
         agrona://stream/bidirectional
         option reader ${agrona:oneToOneReader(targetInput.buffer)}
@@ -41,23 +43,33 @@ read notify HTTP_REQUEST_RECEIVED
 
 write [0x40 0x00 0x00 0x02]
 write ${targetInputId}
+write [0x06 0x00 0x00 0x00]
+write flush
+
+read [0x00 0x00 0x00 0x02]
+read ${targetInputId}
+read [0x00 0x06]
+read "hello "
+
+write [0x40 0x00 0x00 0x02]
+write ${targetInputId}
 write [0x03 0x00 0x00 0x00]
 write flush
 
 read [0x00 0x00 0x00 0x02]
 read ${targetInputId}
 read [0x00 0x03]
-read "hel"
+read "wor"
 
 write [0x40 0x00 0x00 0x02]
 write ${targetInputId}
-write [0x08 0x00 0x00 0x00]
+write [0x02 0x00 0x00 0x00]
 write flush
 
 read [0x00 0x00 0x00 0x02]
 read ${targetInputId}
-read [0x00 0x0b]
-read "lo world"
+read [0x00 0x02]
+read "ld"
 
 read [0x00 0x00 0x00 0x03]
 read ${targetInputId}
