+++ src/test/java/org/reaktivity/specification/nukleus/http/streams/rfc7230/MessageFormatIT.java
@@ -61,31 +61,6 @@ public class MessageFormatIT
 
     @Test
     @Specification({
-        "${streams}/request.with.content.length.and.end.late.target.window/server/source",
-        "${streams}/request.with.content.length.and.end.late.target.window/server/nukleus",
-        "${streams}/request.with.content.length.and.end.late.target.window/server/target" })
-    public void shouldWaitForTargetWindowBeforeProcessingEnd() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${streams}/request.with.content.length.no.target.window/server/source",
-        "${streams}/request.with.content.length.no.target.window/server/nukleus",
-        "${streams}/request.with.content.length.no.target.window/server/target" })
-    public void shouldNotWriteDataToTargetWithoutWindow() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
         "${streams}/request.with.headers/server/source",
         "${streams}/request.with.headers/server/nukleus",
         "${streams}/request.with.headers/server/target" })
@@ -94,83 +69,6 @@ public class MessageFormatIT
         k3po.start();
         k3po.notifyBarrier("ROUTED_INPUT");
         k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${streams}/request.fragmented/server/source",
-        "${streams}/request.fragmented/server/nukleus",
-        "${streams}/request.fragmented/server/target" })
-    public void shouldAcceptFragmentedRequest() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${streams}/request.fragmented.with.content.length/server/source",
-        "${streams}/request.fragmented.with.content.length/server/nukleus",
-        "${streams}/request.fragmented.with.content.length/server/target" })
-    public void shouldAcceptFragmentedRequestWithContentLength() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${streams}/request.headers.too.long/server/source",
-        "${streams}/request.headers.too.long/server/nukleus" })
-    public void shouldRejectRequestExceedingMaximumHeadersSize() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-
-    @Test
-    @Specification({
-        "${streams}/response.fragmented/client/source",
-        "${streams}/response.fragmented/client/nukleus",
-        "${streams}/response.fragmented/client/target" })
-    public void shouldAcceptFragmentedResponse() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-    @Test
-    @Specification({
-        "${streams}/response.fragmented.with.content.length/client/source",
-        "${streams}/response.fragmented.with.content.length/client/nukleus",
-        "${streams}/response.fragmented.with.content.length/client/target" })
-    public void shouldAcceptFragmentedResponseWithContentLength() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${streams}/response.headers.too.long/client/source",
-        "${streams}/response.headers.too.long/client/nukleus",
-        "${streams}/response.headers.too.long/client/target"})
-    public void shouldRejectResponseExceedingMaximumHeadersSize() throws Exception
-    {
-        k3po.start();
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.notifyBarrier("ROUTED_INPUT");
         k3po.finish();
     }
